report(test): split out axe test
